Shared package improvements
