+++ package.json
@@ -34,11 +34,12 @@
   "private": true,
   "workspaces": [
     "vscode",
-    "webview-ui"
+    "webview-ui",
+    "shared"
   ],
   "scripts": {
     "dev": "concurrently \"npm run dev -w vscode\" \"npm run start -w webview-ui\"",
-    "build": "npm run build -w webview-ui && npm run build -w vscode",
+    "build": "npm run build -w shared && npm run build -w webview-ui && npm run build -w vscode ",
     "package": "npm run build && npm run package -w vscode",
     "lint": "eslint .",
     "lint:fix": "eslint . --fix",
