+++ shared/package.json
@@ -1,0 +1,16 @@
+{
+    "name": "@editor-extensions/shared",
+    "version": "1.0.0",
+    "private": true,
+    "main": "./dist/index.js",
+    "types": "./dist/index.d.ts",
+    "scripts": {
+      "build": "tsc -b tsconfig.json",
+      "clean": "rimraf dist",
+      "prebuild": "npm run clean"
+    },
+    "devDependencies": {
+      "rimraf": "^5.0.0",
+      "typescript": "^5.0.0"
+    }
+  }
