+++ tsconfig.json
@@ -1,0 +1,27 @@
+{
+  "compilerOptions": {
+    "target": "ES2020",
+    "module": "commonjs",
+    "strict": true,
+    "esModuleInterop": true,
+    "skipLibCheck": true,
+    "forceConsistentCasingInFileNames": true,
+    "baseUrl": ".",
+    "paths": {
+      "@shared/*": ["shared/src/*"]
+    },
+    "composite": true,
+    "declaration": true
+  },
+  "include": [
+    "shared/src/**/*.ts",
+    "vscode/src/**/*.ts",
+    "webview-ui/src/**/*.ts"
+  ],
+
+  "references": [
+    { "path": "./shared" },
+    { "path": "./vscode" },
+    { "path": "./webview-ui" }
+  ]
+}
