+++ vscode/package.json
@@ -282,6 +282,7 @@
     "webpack-cli": "^5.1.4"
   },
   "dependencies": {
-    "babel-loader": "^9.2.1"
+    "babel-loader": "^9.2.1",
+    "@shared/types": "^1.0.0"
   }
 }
