+++ vscode/src/test/analyzerResults.test.ts
@@ -2,7 +2,7 @@ import * as assert from "assert";
 import * as vscode from "vscode";
 import * as path from "path";
 import { processIncidents, readYamlFile } from "../client/analyzerResults";
-import { RuleSet } from "../shared/types";
+import { RuleSet } from "@shared/types";
 
 suite("Extension Test Suite", () => {
   test("processIncidents should populate diagnostics correctly", () => {
