+++ webview-ui/package.json
@@ -14,7 +14,8 @@
     "@patternfly/react-core": "6.0.0-prerelease.15",
     "@patternfly/react-icons": "^5.4.0",
     "@patternfly/react-table": "^5.4.1",
-    "vscode-webview": "^1.0.1-beta.1"
+    "vscode-webview": "^1.0.1-beta.1",
+    "@shared/types": "^1.0.0"
   },
   "devDependencies": {
     "@types/vscode-webview": "^1.57.5",
