+++ webview-ui/vite.config.ts
@@ -1,16 +1,14 @@
 import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react";
-import { resolve } from "path";
+import path from "path";
 
 export default defineConfig({
   plugins: [react()],
+  resolve: {},
   build: {
     outDir: "build",
     sourcemap: true,
     rollupOptions: {
-      input: {
-        main: resolve(__dirname, "index.html"),
-      },
       output: {
         entryFileNames: `assets/[name].js`,
         chunkFileNames: `assets/[name].js`,
